added spell-helper and a file parser
spell-helper is invoked by prefix:spell-helper {name of spell} if the spell is found, a description is returned. if the spell is not found, it searches by the first 3 letters of given spell name for a partial match, and returns all those spell names. it uses a carefully formatted text file to get spell info, and uses the file parser to retrieve all that data into memory.