finished rule-help, and updated help
rule-help is not fully implemented, and has an easy to use selector that allows users to select a command without typing the full name.
the help function is now fully dynamic, as we add or remove commands, it will update the embed file.
Removed spellFileParser and ruleFileParser, as they were both redundant, especially now with the fileParser method implemented. Added a rule_list.txt file, containing a few rules for demo purposes.